+++ contribution/versions/ezpublish2/trunk/projects/php/ezforum/admin/messageedit.php
@@ -1,6 +1,6 @@
 <?
 /*!
-    $Id: messageedit.php,v 1.3 2000/10/20 13:31:31 ce-cvs Exp $
+    $Id: messageedit.php,v 1.4 2000/10/20 15:28:44 ce-cvs Exp $
 
     Author: Lars Wilhelmsen <lw@ez.no>
     
@@ -14,6 +14,7 @@ $ini = new INIFile( "site.ini" );
 
 $DOC_ROOT = $ini->read_var( "eZForumMain", "DocumentRoot" );
 $Language = $ini->read_var( "eZForumMain", "Language" );
+$error = new INIFIle( "ezforum/admin/intl/" . $Language . "/messageedit.php.ini", false );
 
 include_once( "ezforum/classes/ezforummessage.php" );
 include_once( "ezforum/classes/ezforum.php" );
@@ -21,41 +22,75 @@ include_once( "ezforum/classes/ezforumcategory.php" );
 include_once( "classes/eztemplate.php" );
 include_once( "classes/ezlocale.php" );
 
+require( "ezuser/admin/admincheck.php" );
+
 if ( $Action == "insert" )
 {
+    // Admin does not support insert.
 }
 if ( $Action == "update" )
 {
-    $msg = new eZForumMessage();
-    $msg->get( $MessageID );
-    $msg->setTopic( $Topic );
-    $msg->setBody( $Body );
+    if ( eZPermission::checkPermission( $user, "eZForum", "MessageModify" ) )
+    {
+        if ( $Topic != "" &&
+        $Body != "" )
+        {
+            $msg = new eZForumMessage();
+            $msg->get( $MessageID );
+            $msg->setTopic( $Topic );
+            $msg->setBody( $Body );
 
-    if ( $notice )
-        $msg->enableEmailNotice();
-    else 
-        $msg->disableEmailNotice();
+            if ( $notice )
+                $msg->enableEmailNotice();
+            else 
+                $msg->disableEmailNotice();
 
-    $ForumID = $msg->forumID();
-    $forum = new eZForum( $ForumID );
-    $CategoryID = $forum->categoryID();
-    $msg->store();
-    Header( "Location: /forum/messagelist/$ForumID/" );
-    exit();
+            $ForumID = $msg->forumID();
+            $forum = new eZForum( $ForumID );
+            $CategoryID = $forum->categoryID();
+            $msg->store();
+
+            Header( "Location: /forum/messagelist/$ForumID/" );
+            exit();
+        }
+        else
+        {
+            $error_msg = $error->read_var( "strings", "error_missingdata" );
+        }
+    }
+    else
+    {
+        Header( "Location: /forum/norights" );
+   }
+    
 }
 
 if ( $Action == "delete" )
 {
-    $msg = new eZForumMessage();
-    $msg->get( $MessageID );
-    $msg->delete();
-
-    $ForumID = $msg->forumID();
-    $forum = new eZForum( $ForumID );
-    $CategoryID = $forum->categoryID();
-
-    Header( "Location: /forum/messagelist/$ForumID" );
-    exit();
+    if ( eZPermission::checkPermission( $user, "eZForum", "MessageDelete" ) )
+    {
+        if ( $MessageID != "" )
+        {
+            $msg = new eZForumMessage();
+            $msg->get( $MessageID );
+            $msg->delete();
+            
+            $ForumID = $msg->forumID();
+            $forum = new eZForum( $ForumID );
+            $CategoryID = $forum->categoryID();
+            
+            Header( "Location: /forum/messagelist/$ForumID" );
+            exit();
+        }
+        else
+        {
+            $error_msg = $error->read_var( "strings", "error_missingdata" );
+        }
+    }
+    else
+    {
+        Header( "Location: /forum/norights" );
+    }
 }
 
 $t = new eZTemplate( "ezforum/admin/" . $ini->read_var( "eZForumMain", "TemplateDir" ),
@@ -70,21 +105,50 @@ $headline =  $ini->read_var( "strings", "head_line_insert" );
 $t->set_block( "message_page", "message_edit_tpl", "message_edit" );
 $locale = new eZLocale( $Language );
 
+$t->set_var( "message_topic", "" );
+$t->set_var( "message_postingtime", "" );
+$t->set_var( "message_body", "" );
+$t->set_var( "message_user", "" );
+$t->set_var( "message_id", $MessageID );
+$action_value = "update";
+
+if ( $Action == "new" )
+{
+    if ( !eZPermission::checkPermission( $user, "eZForum", "MessageModify" ) )
+    {
+        Header( "Location: /forum/norights" );
+    }
+
+    $action_value = "insert";
+}
+
+
 if ( $Action == "edit" )
 {
-    $msg = new eZForumMessage();
-    $msg->get( $MessageID );
-    $t->set_var( "message_topic", $msg->topic() );
-    $t->set_var( "message_postingtime", $locale->format( $msg->postingTime() ) );
-    $t->set_var( "message_body", $msg->body() );
-    $user = $msg->user();
-    $t->set_var( "message_user", $user->firstName() . " " . $user->lastName() );
     $ini = new INIFile( $DOC_ROOT . "/admin/" . "intl/" . $Language . "/messageedit.php.ini", false );
     $headline =  $ini->read_var( "strings", "head_line_edit" );
-    $t->set_var( "action_value", "update" );
-    $t->set_var( "message_id", $MessageID );
-    $t->set_var( "forum_id", $msg->forumID() );
+
+    if ( !eZPermission::checkPermission( $user, "eZForum", "MessageModify" ) )
+    {
+        Header( "Location: /forum/norights" );
+    }
+    else
+    {
+        $msg = new eZForumMessage();
+        $msg->get( $MessageID );
+        $t->set_var( "message_topic", $msg->topic() );
+        $t->set_var( "message_postingtime", $locale->format( $msg->postingTime() ) );
+        $t->set_var( "message_body", $msg->body() );
+        $user = $msg->user();
+        $t->set_var( "message_user", $user->firstName() . " " . $user->lastName() );
+        $action_value = "update";
+        $t->set_var( "message_id", $MessageID );
+        $t->set_var( "forum_id", $msg->forumID() );
+    }
 }
+
+$t->set_var( "action_value", $action_value );
+$t->set_var( "error_msg", $error_msg );
 $t->set_var( "category_id", $CategoryID );
 $t->set_var( "headline", $headline );
 $t->pparse( "output", "message_page" );
